refactor: 대타 기능 구현
