+++ src/pages/SchedulePage/DailyWorkerSection/DailyWorkers.tsx
@@ -1,16 +1,12 @@
-import DailyWorkersTemplate from 'components/DailyWorkers/DailyWorkersTemplate';
 import React from 'react';
 import { useAtomValue } from 'jotai';
 import { dateAtom } from '../states';
-import NotFixedDateBox from 'components/DailyWorkers/NotFixedDateBox';
+import { DailyWorkersTable, NotFixedDateBox } from 'components/DailyWorkersTable';
 import { useGetDailyWorkers } from 'hooks/SchedulePage/fetch';
-// import SubmitButton from 'components/@commons/SubmitButton';
-// import { getLoginData } from 'utils/loginDatahandlers';
 
 const DailyWorkers = (): JSX.Element => {
   const selectedDate = useAtomValue(dateAtom);
   const { scheduleResponse } = useGetDailyWorkers();
-  //  const isAdmin = getLoginData().isAdmin;
 
   if (selectedDate.date === '') {
     return <></>;
@@ -22,8 +18,7 @@ const DailyWorkers = (): JSX.Element => {
 
   return (
     <>
-      <DailyWorkersTemplate dailyData={scheduleResponse?.schedule} />
-      {/* {!isAdmin && <SubmitButton>대타를 구하고 싶어요</SubmitButton>} */}
+      <DailyWorkersTable dailyData={scheduleResponse?.schedule} />
     </>
   );
 };
